+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## [Unreleased][]
 
-[Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-aws/compare/0.18.0...HEAD
+[Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-aws/compare/0.19.0...HEAD
+
+## [0.19.0][] - 2021-12-02
+[0.19.0]: https://github.com/chaostoolkit-incubator/chaostoolkit-aws/compare/0.18.0...0.19.0
 
 ### Added
 - workflow to mark issues as `stale` and remove them after 7 days of being `stale`